Move Parser, Storage and Ui to duke directory
