+++ src/main/java/duke/exception/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package ip.exception;
+package duke.exception;
 
 public class EmptyDescriptionException extends Exception {
     public EmptyDescriptionException() {
