+++ src/main/java/duke/exception/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package ip.exception;
+package duke.exception;
 
 public class UnknownCommandException extends Exception{
     public UnknownCommandException() {
