+++ src/main/java/duke/tasks/Todo.java
@@ -1,4 +1,4 @@
-package ip.tasks;
+package duke.tasks;
 
 public class Todo extends Task {
     public Todo(String description) {
